requirements.txt
playsound
pygobject
RPi.GPIO

